Move pagination util functions to pagination module
